+++ resources/views/layouts/master.blade.php
@@ -86,6 +86,10 @@ height="0" width="0" style="display:none;visibility:hidden"></iframe></noscript>
 		</div>
 	@endif
 
+@if (config('app.scena_ai_key'))
+	<scena-ai-popup key-id="{{ config('app.scena_ai_key') }}" id="scena-ai"></scena-ai-popup>
+	<script src="https://widget.scena.ai/app.js"></script>
+@endif
 
 	<!-- Content -->
 	<div id="app">